[account address] Move account address to separate file
